+++ packages/perseus-core/src/parse-perseus-json/perseus-parsers/widgets-map.test.ts
@@ -565,12 +565,13 @@ describe("parseWidgetsMap", () => {
         const widgetsMap: unknown = {
             "numeric-input 1": {
                 type: "numeric-input",
-                version: {major: 0, minor: 0},
+                version: {major: 1, minor: 0},
                 options: {
                     answers: [],
                     labelText: "",
                     size: "",
                     coefficient: false,
+                    textAlign: "left",
                 },
             },
         };