+++ packages/perseus-core/src/utils/extract-perseus-ai-data.test.ts
@@ -828,6 +828,7 @@ describe("injectWidgets", () => {
                     labelText: "Enter a number",
                     size: "normal",
                     coefficient: false,
+                    textAlign: "left",
                 },
             },
             "input-number 1": {
@@ -1020,6 +1021,7 @@ describe("getAnswersFromWidgets", () => {
                 labelText: "Enter a number",
                 size: "normal",
                 coefficient: false,
+                textAlign: "left",
             },
         };
         const answer = getAnswersFromWidgets({"numeric-input 1": widget});